Fixed loggedIn always returning true (almost)
Problem in that a null user was being stored in the session which itself wasn't null, but had null attributes. This would cause a true return for loggedIn Which, I believe, lead in turn to issues #3 and #4 because the user has null IDs, CRSIDs etc. but is still being considered as "logged in".

Signed-off-by: Gideon Farrell <me@gideonfarrell.co.uk>